refactor(tests): move codewhisperer E2E tests
Problem:
Tests in `testInteg/codewhisperer/e2e/` are using
a `skiptTestIfNoValidConn()` workaround to "skip" tests:
https://github.com/aws/aws-toolkit-vscode/blob/6dacf7b372b48b6b85a2b828b182965de34f9ab6/src/testInteg/codewhisperer/e2e/referenceTracker.test.ts#L60-L61

This is fragile and causes problems such as described in #4241.

Solution:
Move tests that require authentication, into the testE2E/ folder which
exists for that purpose.
