+++ src/testE2E/amazonqGumby/transformByQ.test.ts
