+++ src/testE2E/codewhisperer/referenceTracker.test.ts
@@ -4,13 +4,13 @@
  */
 
 import assert from 'assert'
-import * as codewhispererClient from '../../../codewhisperer/client/codewhisperer'
-import { ConfigurationEntry } from '../../../codewhisperer/models/model'
-import { setValidConnection, skiptTestIfNoValidConn } from '../../util/codewhispererUtil'
-import { RecommendationHandler } from '../../../codewhisperer/service/recommendationHandler'
-import { createMockTextEditor, resetCodeWhispererGlobalVariables } from '../../../test/codewhisperer/testUtil'
-import { invokeRecommendation } from '../../../codewhisperer/commands/invokeRecommendation'
-import { session } from '../../../codewhisperer/util/codeWhispererSession'
+import * as codewhispererClient from '../../codewhisperer/client/codewhisperer'
+import { ConfigurationEntry } from '../../codewhisperer/models/model'
+import { setValidConnection, skiptTestIfNoValidConn } from '../util/codewhispererUtil'
+import { RecommendationHandler } from '../../codewhisperer/service/recommendationHandler'
+import { createMockTextEditor, resetCodeWhispererGlobalVariables } from '../../test/codewhisperer/testUtil'
+import { invokeRecommendation } from '../../codewhisperer/commands/invokeRecommendation'
+import { session } from '../../codewhisperer/util/codeWhispererSession'
 
 /*
 New model deployment may impact references returned.
