+++ src/testE2E/util/amazonQUtil.ts
